+++ packages/mintlify/package.json
@@ -28,13 +28,15 @@
     "node-html-markdown": "^1.2.0",
     "open": "^8.4.0",
     "puppeteer": "^17.1.3",
-    "shelljs": "^0.8.5"
+    "shelljs": "^0.8.5",
+    "yargs": "^17.6.0"
   },
   "devDependencies": {
     "@types/fs-extra": "^9.0.13",
     "@types/inquirer": "^9.0.1",
     "@types/node": "^18.7.13",
     "@types/shelljs": "^0.8.11",
+    "@types/yargs": "^17.0.13",
     "typescript": "^4.8.2"
   }
 }
