+++ packages/mintlify/src/init-command/templates.ts
